+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from "../components/index"
+import Layout from "../components/layout"
 import { Link, graphql } from 'gatsby'
 import Helmet from 'react-helmet'
